ai_expense_tracker/requirements.txt
streamlit
google-generativeai
pandas
plotly
python-dotenv
watchdog
